refactor amp/bento-date-display
